BitMEX.API

Description

Operations

APIKey

aPIKeyDisable

aPIKeyDisable

Arguments

:: Consumes APIKeyDisable contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> ApiKeyId
"apiKeyId" - API Key ID (public component).

-> BitMEXRequest APIKeyDisable contentType APIKey accept

POST /apiKey/disable

Disable an API Key.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data APIKeyDisable

aPIKeyEnable

aPIKeyEnable

Arguments

:: Consumes APIKeyEnable contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> ApiKeyId
"apiKeyId" - API Key ID (public component).

-> BitMEXRequest APIKeyEnable contentType APIKey accept

POST /apiKey/enable

Enable an API Key.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data APIKeyEnable

aPIKeyGet

aPIKeyGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest APIKeyGet MimeNoContent [APIKey] accept

GET /apiKey

Get your API Keys.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data APIKeyGet

aPIKeyNew

aPIKeyNew

Arguments

:: Consumes APIKeyNew contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest APIKeyNew contentType APIKey accept

POST /apiKey

Create a new API Key.

API Keys can also be created via this Python script See the API Key Documentation for more information on capabilities.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data APIKeyNew

aPIKeyRemove

aPIKeyRemove

Arguments

:: Consumes APIKeyRemove contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> ApiKeyId
"apiKeyId" - API Key ID (public component).

-> BitMEXRequest APIKeyRemove contentType InlineResponse200 accept

DELETE /apiKey

Remove an API Key.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data APIKeyRemove

Announcement

announcementGet

announcementGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest AnnouncementGet MimeNoContent [Announcement] accept

GET /announcement

Get site announcements.

data AnnouncementGet

announcementGetUrgent

announcementGetUrgent

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest AnnouncementGetUrgent MimeNoContent [Announcement] accept

GET /announcement/urgent

Get urgent (banner) announcements.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data AnnouncementGetUrgent

Chat

chatGet

chatGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest ChatGet MimeNoContent [Chat] accept

GET /chat

Get chat messages.

data ChatGet

chatGetChannels

chatGetChannels

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest ChatGetChannels MimeNoContent [ChatChannels] accept

GET /chat/channels

Get available channels.

data ChatGetChannels

chatGetConnected

chatGetConnected

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest ChatGetConnected MimeNoContent ConnectedUsers accept

GET /chat/connected

Get connected users.

Returns an array with browser users in the first position and API users (bots) in the second position.

data ChatGetConnected

chatNew

chatNew

Arguments

:: Consumes ChatNew contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Message
"message"

-> BitMEXRequest ChatNew contentType Chat accept

POST /chat

Send a chat message.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data ChatNew

Execution

executionGet

executionGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest ExecutionGet MimeNoContent [Execution] accept

GET /execution

Get all raw executions for your account.

This returns all raw transactions, which includes order opening and cancelation, and order status changes. It can be quite noisy. More focused information is available at `/execution/tradeHistory`. You may also use the filter param to target your query. Specify an array as a filter value, such as `{"execType": ["Settlement", "Trade"]}` to filter on multiple values. See the FIX Spec for explanations of these fields.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data ExecutionGet

executionGetTradeHistory

executionGetTradeHistory

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest ExecutionGetTradeHistory MimeNoContent [Execution] accept

GET /execution/tradeHistory

Get all balance-affecting executions. This includes each trade, insurance charge, and settlement.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data ExecutionGetTradeHistory

Funding

fundingGet

fundingGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest FundingGet MimeNoContent [Funding] accept

GET /funding

Get funding history.

data FundingGet

Instrument

instrumentGet

instrumentGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest InstrumentGet MimeNoContent [Instrument] accept

GET /instrument

Get instruments.

This returns all instruments and indices, including those that have settled or are unlisted. Use this endpoint if you want to query for individual instruments or use a complex filter. Use `/instrument/active` to return active instruments, or use a filter like `{"state": "Open"}`.

data InstrumentGet

instrumentGetActive

instrumentGetActive

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest InstrumentGetActive MimeNoContent [Instrument] accept

GET /instrument/active

Get all active instruments and instruments that have expired in <24hrs.

data InstrumentGetActive

instrumentGetActiveAndIndices

instrumentGetActiveAndIndices

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest InstrumentGetActiveAndIndices MimeNoContent [Instrument] accept

GET /instrument/activeAndIndices

Helper method. Gets all active instruments and all indices. This is a join of the result of indices and active.

data InstrumentGetActiveAndIndices

instrumentGetActiveIntervals

instrumentGetActiveIntervals

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest InstrumentGetActiveIntervals MimeNoContent InstrumentInterval accept

GET /instrument/activeIntervals

Return all active contract series and interval pairs.

This endpoint is useful for determining which pairs are live. It returns two arrays of strings. The first is intervals, such as `["BVOL:daily", "BVOL:weekly", "XBU:daily", "XBU:monthly", ...]`. These identifiers are usable in any query's symbol param. The second array is the current resolution of these intervals. Results are mapped at the same index.

data InstrumentGetActiveIntervals

instrumentGetCompositeIndex

instrumentGetCompositeIndex

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest InstrumentGetCompositeIndex MimeNoContent [IndexComposite] accept

GET /instrument/compositeIndex

Show constituent parts of an index.

Composite indices are built from multiple external price sources. Use this endpoint to get the underlying prices of an index. For example, send a symbol of `.XBT` to get the ticks and weights of the constituent exchanges that build the ".XBT" index. A tick with reference `"BMI"` and weight null is the composite index tick.

data InstrumentGetCompositeIndex

instrumentGetIndices

instrumentGetIndices

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest InstrumentGetIndices MimeNoContent [Instrument] accept

GET /instrument/indices

Get all price indices.

data InstrumentGetIndices

Insurance

insuranceGet

insuranceGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest InsuranceGet MimeNoContent [Insurance] accept

GET /insurance

Get insurance fund history.

data InsuranceGet

Leaderboard

leaderboardGet

leaderboardGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest LeaderboardGet MimeNoContent [Leaderboard] accept

GET /leaderboard

Get current leaderboard.

data LeaderboardGet

Liquidation

liquidationGet

liquidationGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest LiquidationGet MimeNoContent [Liquidation] accept

GET /liquidation

Get liquidation orders.

data LiquidationGet

Notification

notificationGet

notificationGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest NotificationGet MimeNoContent [Notification] accept

GET /notification

Get your current notifications.

This is an upcoming feature and currently does not return data.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data NotificationGet

Order

orderAmend

orderAmend

Arguments

:: Consumes OrderAmend contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest OrderAmend contentType Order accept

PUT /order

Amend the quantity or price of an open order.

Send an orderID or origClOrdID to identify the order you wish to amend. Both order quantity and price can be amended. Only one qty field can be used to amend. Use the leavesQty field to specify how much of the order you wish to remain open. This can be useful if you want to adjust your position's delta by a certain amount, regardless of how much of the order has already filled. Use the simpleOrderQty and simpleLeavesQty fields to specify order size in Bitcoin, rather than contracts. These fields will round up to the nearest contract. Like order placement, amending can be done in bulk. Simply send a request to `PUT apiv1orderbulk` with a JSON body of the shape: `{"orders": [{...}, {...}]}`, each object containing the fields used in this endpoint.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderAmend

orderAmendBulk

orderAmendBulk

Arguments

:: Consumes OrderAmendBulk contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest OrderAmendBulk contentType [Order] accept

PUT /order/bulk

Amend multiple orders for the same symbol.

Similar to POST amend, but with multiple orders. `applicationjson` only. Ratelimited at 50%.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderAmendBulk

orderCancel

orderCancel

Arguments

:: Consumes OrderCancel contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest OrderCancel contentType [Order] accept

DELETE /order

Cancel order(s). Send multiple order IDs to cancel in bulk.

Either an orderID or a clOrdID must be provided.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderCancel

orderCancelAll

orderCancelAll

Arguments

:: Consumes OrderCancelAll contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest OrderCancelAll contentType Value accept

DELETE /order/all

Cancels all of your orders.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderCancelAll

orderCancelAllAfter

orderCancelAllAfter

Arguments

:: Consumes OrderCancelAllAfter contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Timeout
"timeout" - Timeout in ms. Set to 0 to cancel this timer.

-> BitMEXRequest OrderCancelAllAfter contentType Value accept

POST /order/cancelAllAfter

Automatically cancel all your orders after a specified timeout.

Useful as a dead-man's switch to ensure your orders are canceled in case of an outage. If called repeatedly, the existing offset will be canceled and a new one will be inserted in its place. Example usage: call this route at 15s intervals with an offset of 60000 (60s). If this route is not called within 60 seconds, all your orders will be automatically canceled. This is also available via WebSocket.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderCancelAllAfter

orderClosePosition

orderClosePosition

Arguments

:: Consumes OrderClosePosition contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Symbol of position to close.

-> BitMEXRequest OrderClosePosition contentType Order accept

POST /order/closePosition

Close a position. [Deprecated, use POST /order with execInst: Close]

If no price is specified, a market order will be submitted to close the whole of your position. This will also close all other open orders in this symbol.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderClosePosition

orderGetOrders

orderGetOrders

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest OrderGetOrders MimeNoContent [Order] accept

GET /order

Get your orders.

To get open orders only, send {"open": true} in the filter param. See href="http://www.onixs.biz/fix-dictionary/5.0.SP2/msgType_D_68.html"the FIX Spec/a for explanations of these fields.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderGetOrders

orderNew

orderNew

Arguments

:: Consumes OrderNew contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Instrument symbol. e.g. XBTUSD.

-> BitMEXRequest OrderNew contentType Order accept

POST /order

Create a new order.

#############!OrderOrder_newBulk) if you need to place multiple orders at the same time. Bulk orders require fewer risk checks in the trading engine and thus are ratelimited at **110** the normal rate. You can also improve your reactivity to market movements while staying under your ratelimit by using the [Amend](#!OrderOrder_amend) and [Amend Bulk](#!OrderOrder_amendBulk) endpoints. This allows you to stay in the market and avoids the cancelreplace cycle. #### Tracking Your Orders If you want to keep track of order IDs yourself, set a unique clOrdID per order. This clOrdID will come back as a property on the order and any related executions (including on the WebSocket), and can be used to get or cancel the order. Max length is 36 characters.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderNew

orderNewBulk

orderNewBulk

Arguments

:: Consumes OrderNewBulk contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest OrderNewBulk contentType [Order] accept

POST /order/bulk

Create multiple new orders for the same symbol.

This endpoint is used for placing bulk orders. Valid order types are Market, Limit, Stop, StopLimit, MarketIfTouched, LimitIfTouched, MarketWithLeftOverAsLimit, and Pegged. Each individual order object in the array should have the same properties as an individual POST order call. This endpoint is much faster for getting many orders into the book at once. Because it reduces load on BitMEX systems, this endpoint is ratelimited at `ceil(0.1 * orders)`. Submitting 10 orders via a bulk order call will only count as 1 request, 15 as 2, 32 as 4, and so on. For now, only `applicationjson` is supported on this endpoint.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data OrderNewBulk

OrderBook

orderBookGet

orderBookGet

Arguments

:: Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Instrument symbol. Send a series (e.g. XBT) to get data for the nearest contract in that series.

-> BitMEXRequest OrderBookGet MimeNoContent [OrderBook] accept

GET /orderBook

Get current orderbook [deprecated, use orderBookL2].

data OrderBookGet

orderBookGetL2

orderBookGetL2

Arguments

:: Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Instrument symbol. Send a series (e.g. XBT) to get data for the nearest contract in that series.

-> BitMEXRequest OrderBookGetL2 MimeNoContent [OrderBookL2] accept

GET /orderBook/L2

Get current orderbook in vertical format.

data OrderBookGetL2

Position

positionGet

positionGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest PositionGet MimeNoContent [Position] accept

GET /position

Get your positions.

See href="http://www.onixs.biz/fix-dictionary/5.0.SP2/msgType_AP_6580.html"the FIX Spec/a for explanations of these fields.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data PositionGet

positionIsolateMargin

positionIsolateMargin

Arguments

:: Consumes PositionIsolateMargin contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Position symbol to isolate.

-> BitMEXRequest PositionIsolateMargin contentType Position accept

POST /position/isolate

Enable isolated margin or cross margin per-position.

Users can switch isolate margin per-position. This function allows switching margin isolation (aka fixed margin) on and off.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data PositionIsolateMargin

positionTransferIsolatedMargin

positionTransferIsolatedMargin

Arguments

:: Consumes PositionTransferIsolatedMargin contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Symbol of position to isolate.

-> Amount
"amount" - Amount to transfer, in Satoshis. May be negative.

-> BitMEXRequest PositionTransferIsolatedMargin contentType Position accept

POST /position/transferMargin

Transfer equity in or out of a position.

When margin is isolated on a position, use this function to add or remove margin from the position. Note that you cannot remove margin below the initial margin threshold.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data PositionTransferIsolatedMargin

positionUpdateLeverage

positionUpdateLeverage

Arguments

:: Consumes PositionUpdateLeverage contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Symbol of position to adjust.

-> Leverage
"leverage" - Leverage value. Send a number between 0.01 and 100 to enable isolated margin with a fixed leverage. Send 0 to enable cross margin.

-> BitMEXRequest PositionUpdateLeverage contentType Position accept

POST /position/leverage

Choose leverage for a position.

Users can choose an isolated leverage. This will automatically enable isolated margin.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data PositionUpdateLeverage

positionUpdateRiskLimit

positionUpdateRiskLimit

Arguments

:: Consumes PositionUpdateRiskLimit contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Symbol
"symbol" - Symbol of position to isolate.

-> RiskLimit
"riskLimit" - New Risk Limit, in Satoshis.

-> BitMEXRequest PositionUpdateRiskLimit contentType Position accept

POST /position/riskLimit

Update your risk limit.

Risk Limits limit the size of positions you can trade at various margin levels. Larger positions require more margin. Please see the Risk Limit documentation for more details.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data PositionUpdateRiskLimit

Quote

quoteGet

quoteGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest QuoteGet MimeNoContent [Quote] accept

GET /quote

Get Quotes.

data QuoteGet

quoteGetBucketed

quoteGetBucketed

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest QuoteGetBucketed MimeNoContent [Quote] accept

GET /quote/bucketed

Get previous quotes in time buckets.

data QuoteGetBucketed

Schema

schemaGet

schemaGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest SchemaGet MimeNoContent Value accept

GET /schema

Get model schemata for data objects returned by this API.

data SchemaGet

schemaWebsocketHelp

schemaWebsocketHelp

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest SchemaWebsocketHelp MimeNoContent Value accept

GET /schema/websocketHelp

Returns help text & subject list for websocket usage.

data SchemaWebsocketHelp

Settlement

settlementGet

settlementGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest SettlementGet MimeNoContent [Settlement] accept

GET /settlement

Get settlement history.

data SettlementGet

Stats

statsGet

statsGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest StatsGet MimeNoContent [Stats] accept

GET /stats

Get exchange-wide and per-series turnover and volume statistics.

data StatsGet

statsHistory2

statsHistory2

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest StatsHistory2 MimeNoContent [StatsHistory] accept

GET /stats/history

Get historical exchange-wide and per-series turnover and volume statistics.

data StatsHistory2

statsHistoryUSD

statsHistoryUSD

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest StatsHistoryUSD MimeNoContent [StatsUSD] accept

GET /stats/historyUSD

Get a summary of exchange statistics in USD.

data StatsHistoryUSD

Trade

tradeGet

tradeGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest TradeGet MimeNoContent [Trade] accept

GET /trade

Get Trades.

Please note that indices (symbols starting with .) post trades at intervals to the trade feed. These have a size of 0 and are used only to indicate a changing price. See the FIX Spec for explanations of these fields.

data TradeGet

tradeGetBucketed

tradeGetBucketed

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest TradeGetBucketed MimeNoContent [TradeBin] accept

GET /trade/bucketed

Get previous trades in time buckets.

data TradeGetBucketed

User

userCancelWithdrawal

userCancelWithdrawal

Arguments

:: Consumes UserCancelWithdrawal contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Token
"token"

-> BitMEXRequest UserCancelWithdrawal contentType Transaction accept

POST /user/cancelWithdrawal

Cancel a withdrawal.

data UserCancelWithdrawal

userCheckReferralCode

userCheckReferralCode

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserCheckReferralCode MimeNoContent Double accept

GET /user/checkReferralCode

Check if a referral code is valid.

If the code is valid, responds with the referral code's discount (e.g. `0.1` for 10%). Otherwise, will return a 404.

data UserCheckReferralCode

userConfirm

userConfirm

Arguments

:: Consumes UserConfirm contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Token
"token"

-> BitMEXRequest UserConfirm contentType AccessToken accept

POST /user/confirmEmail

Confirm your email address with a token.

data UserConfirm

userConfirmEnableTFA

userConfirmEnableTFA

Arguments

:: Consumes UserConfirmEnableTFA contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Token
"token" - Token from your selected TFA type.

-> BitMEXRequest UserConfirmEnableTFA contentType Bool accept

POST /user/confirmEnableTFA

Confirm two-factor auth for this account. If using a Yubikey, simply send a token to this endpoint.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserConfirmEnableTFA

userConfirmWithdrawal

userConfirmWithdrawal

Arguments

:: Consumes UserConfirmWithdrawal contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Token
"token"

-> BitMEXRequest UserConfirmWithdrawal contentType Transaction accept

POST /user/confirmWithdrawal

Confirm a withdrawal.

data UserConfirmWithdrawal

userDisableTFA

userDisableTFA

Arguments

:: Consumes UserDisableTFA contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Token
"token" - Token from your selected TFA type.

-> BitMEXRequest UserDisableTFA contentType Bool accept

POST /user/disableTFA

Disable two-factor auth for this account.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserDisableTFA

userGet

userGet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGet MimeNoContent User accept

GET /user

Get your user model.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGet

userGetAffiliateStatus

userGetAffiliateStatus

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGetAffiliateStatus MimeNoContent Affiliate accept

GET /user/affiliateStatus

Get your current affiliate/referral status.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGetAffiliateStatus

userGetCommission

userGetCommission

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGetCommission MimeNoContent [UserCommission] accept

GET /user/commission

Get your account's commission status.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGetCommission

userGetDepositAddress

userGetDepositAddress

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGetDepositAddress MimeNoContent Text accept

GET /user/depositAddress

Get a deposit address.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGetDepositAddress

userGetMargin

userGetMargin

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGetMargin MimeNoContent Margin accept

GET /user/margin

Get your account's margin status. Send a currency of "all" to receive an array of all supported currencies.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGetMargin

userGetWallet

userGetWallet

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGetWallet MimeNoContent Wallet accept

GET /user/wallet

Get your current wallet information.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGetWallet

userGetWalletHistory

userGetWalletHistory

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGetWalletHistory MimeNoContent [Transaction] accept

GET /user/walletHistory

Get a history of all of your wallet transactions (deposits, withdrawals, PNL).

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGetWalletHistory

userGetWalletSummary

userGetWalletSummary

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserGetWalletSummary MimeNoContent [Transaction] accept

GET /user/walletSummary

Get a summary of all of your wallet transactions (deposits, withdrawals, PNL).

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserGetWalletSummary

userLogout

userLogout

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserLogout MimeNoContent res accept

POST /user/logout

Log out of BitMEX.

Note: Has Produces instances, but no response schema

data UserLogout

userLogoutAll

userLogoutAll

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserLogoutAll MimeNoContent Double accept

POST /user/logoutAll

Log all systems out of BitMEX. This will revoke all of your account's access tokens, logging you out on all devices.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserLogoutAll

userMinWithdrawalFee

userMinWithdrawalFee

Arguments

:: Accept accept
request accept (MimeType)

-> BitMEXRequest UserMinWithdrawalFee MimeNoContent Value accept

GET /user/minWithdrawalFee

Get the minimum withdrawal fee for a currency.

This is changed based on network conditions to ensure timely withdrawals. During network congestion, this may be high. The fee is returned in the same currency.

data UserMinWithdrawalFee

userRequestEnableTFA

userRequestEnableTFA

Arguments

:: Consumes UserRequestEnableTFA contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest UserRequestEnableTFA contentType Bool accept

POST /user/requestEnableTFA

Get secret key for setting up two-factor auth.

Use /confirmEnableTFA directly for Yubikeys. This fails if TFA is already enabled.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserRequestEnableTFA

userRequestWithdrawal

userRequestWithdrawal

Arguments

:: Consumes UserRequestWithdrawal contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Currency
"currency" - Currency you're withdrawing. Options: XBt

-> Amount
"amount" - Amount of withdrawal currency.

-> Address
"address" - Destination Address.

-> BitMEXRequest UserRequestWithdrawal contentType Transaction accept

POST /user/requestWithdrawal

Request a withdrawal to an external wallet.

This will send a confirmation email to the email address on record, unless requested via an API Key with the withdraw permission.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserRequestWithdrawal

userSavePreferences

userSavePreferences

Arguments

:: Consumes UserSavePreferences contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Prefs
"prefs"

-> BitMEXRequest UserSavePreferences contentType User accept

POST /user/preferences

Save user preferences.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserSavePreferences

userUpdate

userUpdate

Arguments

:: Consumes UserUpdate contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> BitMEXRequest UserUpdate contentType User accept

PUT /user

Update your password, name, and other attributes.

AuthMethod: AuthApiKeyApiKey, AuthApiKeyApiNonce, AuthApiKeyApiSignature

data UserUpdate

Parameter newtypes

newtype Account

Constructors

Account

Fields

unAccount :: Double

newtype Address

Constructors

Address

Fields

unAddress :: Text

newtype Amount

Constructors

Amount

Fields

unAmount :: Double

newtype ApiKeyId

Constructors

ApiKeyId

Fields

unApiKeyId :: Text

newtype BinSize

Constructors

BinSize

Fields

unBinSize :: Text

newtype ChannelId

Constructors

ChannelId

Fields

unChannelId :: Double

newtype Cidr

Constructors

Cidr

Fields

unCidr :: Text

newtype ClOrdId

Constructors

ClOrdId

Fields

unClOrdId :: Text

newtype ClOrdLinkId

Constructors

ClOrdLinkId

Fields

unClOrdLinkId :: Text

newtype Columns

Constructors

Columns

Fields

unColumns :: Text

newtype ContingencyType

Constructors

ContingencyType

Fields

unContingencyType :: Text

newtype Count

Constructors

Count

Fields

unCount :: Double

newtype Country

Constructors

Country

Fields

unCountry :: Text

newtype Currency

Constructors

Currency

Fields

unCurrency :: Text

newtype Depth

Constructors

Depth

Fields

unDepth :: Double

newtype DisplayQty

Constructors

DisplayQty

Fields

unDisplayQty :: Double

newtype Enabled

Constructors

Enabled

Fields

unEnabled :: Bool

newtype EndTime

Constructors

EndTime

Fields

unEndTime :: DateTime

newtype ExecInst

Constructors

ExecInst

Fields

unExecInst :: Text

newtype Fee

Constructors

Fee

Fields

unFee :: Double

newtype Filter

Constructors

Filter

Fields

unFilter :: Text

newtype Firstname

Constructors

Firstname

Fields

unFirstname :: Text

newtype Lastname

Constructors

Lastname

Fields

unLastname :: Text

newtype LeavesQty

Constructors

LeavesQty

Fields

unLeavesQty :: Double

newtype Leverage

Constructors

Leverage

Fields

unLeverage :: Double

newtype Message

Constructors

Message

Fields

unMessage :: Text

newtype Method

Constructors

Method

Fields

unMethod :: Text

newtype Model

Constructors

Model

Fields

unModel :: Text

newtype Name

Constructors

Name

Fields

unName :: Text

newtype NewPassword

Constructors

NewPassword

Fields

unNewPassword :: Text

newtype NewPasswordConfirm

Constructors

NewPasswordConfirm

Fields

unNewPasswordConfirm :: Text

newtype OldPassword

Constructors

OldPassword

Fields

unOldPassword :: Text

newtype OrdType

Constructors

OrdType

Fields

unOrdType :: Text

newtype OrderId

Constructors

OrderId

Fields

unOrderId :: Text

newtype OrderQty

Constructors

OrderQty

Fields

unOrderQty :: Double

newtype Orders

Constructors

Orders

Fields

unOrders :: Text

newtype OrigClOrdId

Constructors

OrigClOrdId

Fields

unOrigClOrdId :: Text

newtype OtpToken

Constructors

OtpToken

Fields

unOtpToken :: Text

newtype Overwrite

Constructors

Overwrite

Fields

unOverwrite :: Bool

newtype ParamText

Constructors

ParamText

Fields

unParamText :: Text

newtype ParamType

Constructors

ParamType

Fields

unParamType :: Text

newtype Partial

Constructors

Partial

Fields

unPartial :: Bool

newtype PegOffsetValue

Constructors

PegOffsetValue

Fields

unPegOffsetValue :: Double

newtype PegPriceType

Constructors

PegPriceType

Fields

unPegPriceType :: Text

newtype Permissions

Constructors

Permissions

Fields

unPermissions :: Text

newtype PgpPubKey

Constructors

PgpPubKey

Fields

unPgpPubKey :: Text

newtype Prefs

Constructors

Prefs

Fields

unPrefs :: Text

newtype Price

Constructors

Price

Fields

unPrice :: Double

newtype Quantity

Constructors

Quantity

Fields

unQuantity :: Double

newtype ReferralCode

Constructors

ReferralCode

Fields

unReferralCode :: Text

newtype Reverse

Constructors

Reverse

Fields

unReverse :: Bool

newtype RiskLimit

Constructors

RiskLimit

Fields

unRiskLimit :: Double

newtype Side

Constructors

Side

Fields

unSide :: Text

newtype SimpleLeavesQty

Constructors

SimpleLeavesQty

Fields

unSimpleLeavesQty :: Double

newtype SimpleOrderQty

Constructors

SimpleOrderQty

Fields

unSimpleOrderQty :: Double

newtype Start

Constructors

Start

Fields

unStart :: Double

newtype StartTime

Constructors

StartTime

Fields

unStartTime :: DateTime

newtype StopPrice

Constructors

StopPrice

Fields

unStopPrice :: Double

newtype StopPx

Constructors

StopPx

Fields

unStopPx :: Double

newtype Symbol

Constructors

Symbol

Fields

unSymbol :: Text

newtype TimeInForce

Constructors

TimeInForce

Fields

unTimeInForce :: Text

newtype Timeout

Constructors

Timeout

Fields

unTimeout :: Double

newtype Token

Constructors

Token

Fields

unToken :: Text

newtype Username

Constructors

Username

Fields

unUsername :: Text

Auth Methods

AuthApiKeyApiKey

data AuthApiKeyApiKey

Constructors

AuthApiKeyApiKey Text
secret

AuthApiKeyApiNonce

data AuthApiKeyApiNonce

Constructors

AuthApiKeyApiNonce Text
secret

AuthApiKeyApiSignature

data AuthApiKeyApiSignature

Constructors

AuthApiKeyApiSignature Text
secret

Custom Mime Types

MimeJavascript

data MimeJavascript

Constructors

MimeJavascript

MimeTextjavascript

data MimeTextjavascript

Constructors

MimeTextjavascript

MimeTextxml

data MimeTextxml

Constructors

MimeTextxml