BitMEX.Core

Description

BitMEXConfig

data BitMEXConfig

Constructors

BitMEXConfig

Fields

configHost :: ByteString

host supplied in the Request

configUserAgent :: Text

user-agent supplied in the Request

configLogExecWithContext :: LogExecWithContext

Run a block using a Logger instance

configLogContext :: LogContext

Configures the logger

configAuthMethods :: [AnyAuthMethod]

List of configured auth methods

configValidateAuthMethods :: Bool

throw exceptions if auth methods are not configured

newConfig :: IO BitMEXConfig

constructs a default BitMEXConfig

configHost:

https://localhost/api/v1

configUserAgent:

"swagger-haskell-http-client/1.0.0"

addAuthMethod :: AuthMethod auth => BitMEXConfig -> auth -> BitMEXConfig

updates config use AuthMethod on matching requests

withStdoutLogging :: BitMEXConfig -> IO BitMEXConfig

updates the config to use stdout logging

withStderrLogging :: BitMEXConfig -> IO BitMEXConfig

updates the config to use stderr logging

withNoLogging :: BitMEXConfig -> BitMEXConfig

updates the config to disable logging

BitMEXRequest

data BitMEXRequest req contentType res accept

Represents a request.

Type Variables:

req - request operation

contentType - MimeType associated with request body

res - response model

accept - MimeType associated with response body

Constructors

BitMEXRequest

Fields

rMethod :: Method

Method of BitMEXRequest

rUrlPath :: [ByteString]

Endpoint of BitMEXRequest

rParams :: Params

params of BitMEXRequest

rAuthTypes :: [TypeRep]

types of auth methods

rMethodL :: Lens_' (BitMEXRequest req contentType res accept) Method

rMethod Lens

rUrlPathL :: Lens_' (BitMEXRequest req contentType res accept) [ByteString]

rUrlPath Lens

rParamsL :: Lens_' (BitMEXRequest req contentType res accept) Params

rParams Lens

rAuthTypesL :: Lens_' (BitMEXRequest req contentType res accept) [TypeRep]

rParams Lens

HasBodyParam

class HasBodyParam req param where

Designates the body parameter of a request

Methods

setBodyParam :: forall contentType res accept. (Consumes req contentType, MimeRender contentType param) => BitMEXRequest req contentType res accept -> param -> BitMEXRequest req contentType res accept

HasOptionalParam

class HasOptionalParam req param where

Designates the optional parameters of a request

Minimal complete definition

applyOptionalParam | (-&-)

Methods

applyOptionalParam :: BitMEXRequest req contentType res accept -> param -> BitMEXRequest req contentType res accept

Apply an optional parameter to a request

(-&-) :: BitMEXRequest req contentType res accept -> param -> BitMEXRequest req contentType res accept infixl 2

infix operator / alias for addOptionalParam

data Params

Request Params

Constructors

Params

Fields

paramsQuery :: Query

paramsHeaders :: RequestHeaders

paramsBody :: ParamBody

paramsQueryL :: Lens_' Params Query

paramsQuery Lens

paramsHeadersL :: Lens_' Params RequestHeaders

paramsHeaders Lens

paramsBodyL :: Lens_' Params ParamBody

paramsBody Lens

data ParamBody

Request Body

Constructors

ParamBodyNone

ParamBodyB ByteString

ParamBodyBL ByteString

ParamBodyFormUrlEncoded Form

ParamBodyMultipartFormData [Part]

BitMEXRequest Utils

_mkRequest

Arguments

:: Method
Method

-> [ByteString]
Endpoint

-> BitMEXRequest req contentType res accept
req: Request Type, res: Response Type

_mkParams :: Params

setHeader :: BitMEXRequest req contentType res accept -> [Header] -> BitMEXRequest req contentType res accept

removeHeader :: BitMEXRequest req contentType res accept -> [HeaderName] -> BitMEXRequest req contentType res accept

_setContentTypeHeader :: forall req contentType res accept. MimeType contentType => BitMEXRequest req contentType res accept -> BitMEXRequest req contentType res accept

_setAcceptHeader :: forall req contentType res accept. MimeType accept => BitMEXRequest req contentType res accept -> BitMEXRequest req contentType res accept

setQuery :: BitMEXRequest req contentType res accept -> [QueryItem] -> BitMEXRequest req contentType res accept

addForm :: BitMEXRequest req contentType res accept -> Form -> BitMEXRequest req contentType res accept

_addMultiFormPart :: BitMEXRequest req contentType res accept -> Part -> BitMEXRequest req contentType res accept

_setBodyBS :: BitMEXRequest req contentType res accept -> ByteString -> BitMEXRequest req contentType res accept

_setBodyLBS :: BitMEXRequest req contentType res accept -> ByteString -> BitMEXRequest req contentType res accept

_hasAuthType :: AuthMethod authMethod => BitMEXRequest req contentType res accept -> Proxy authMethod -> BitMEXRequest req contentType res accept

Params Utils

toPath :: ToHttpApiData a => a -> ByteString

toHeader :: ToHttpApiData a => (HeaderName, a) -> [Header]

toForm :: ToHttpApiData v => (ByteString, v) -> Form

toQuery :: ToHttpApiData a => (ByteString, Maybe a) -> [QueryItem]

Swagger CollectionFormat Utils

data CollectionFormat

Determines the format of the array if type array is used.

Constructors

CommaSeparated
CSV format for multiple parameters.

SpaceSeparated
Also called SSV

TabSeparated
Also called TSV

PipeSeparated
`value1|value2|value2`

MultiParamArray
Using multiple GET parameters, e.g. `foo=bar&foo=baz`. This is valid only for parameters in "query" (Query) or "formData" (Form)

toHeaderColl :: ToHttpApiData a => CollectionFormat -> (HeaderName, [a]) -> [Header]

toFormColl :: ToHttpApiData v => CollectionFormat -> (ByteString, [v]) -> Form

toQueryColl :: ToHttpApiData a => CollectionFormat -> (ByteString, Maybe [a]) -> Query

_toColl :: Traversable f => CollectionFormat -> (f a -> [(b, ByteString)]) -> f [a] -> [(b, ByteString)]

_toCollA :: (Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t ByteString)]) -> f (t [a]) -> [(b, t ByteString)]

_toCollA' :: (Monoid c, Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t c)]) -> (Char -> c) -> f (t [a]) -> [(b, t c)]

AuthMethods

class Typeable a => AuthMethod a where

Provides a method to apply auth methods to requests

Minimal complete definition

applyAuthMethod

Methods

applyAuthMethod :: BitMEXConfig -> a -> BitMEXRequest req contentType res accept -> IO (BitMEXRequest req contentType res accept)

data AnyAuthMethod

An existential wrapper for any AuthMethod

Constructors

AuthMethod a => AnyAuthMethod a

data AuthMethodException

indicates exceptions related to AuthMethods

Constructors

AuthMethodException String

_applyAuthMethods :: BitMEXRequest req contentType res accept -> BitMEXConfig -> IO (BitMEXRequest req contentType res accept)

apply all matching AuthMethods in config to request

Utils

_omitNulls :: [(Text, Value)] -> Value

Removes Null fields. (OpenAPI-Specification 2.0 does not allow Null in JSON)

_toFormItem :: (ToHttpApiData a, Functor f) => t -> f a -> f (t, [Text])

Encodes fields using WH.toQueryParam

_emptyToNothing :: Maybe String -> Maybe String

Collapse (Just "") to Nothing

_memptyToNothing :: (Monoid a, Eq a) => Maybe a -> Maybe a

Collapse (Just mempty) to Nothing

DateTime Formatting

newtype DateTime

Constructors

DateTime

Fields

unDateTime :: UTCTime

_readDateTime :: (ParseTime t, Monad m, Alternative m) => String -> m t

_parseISO8601

_showDateTime :: (t ~ UTCTime, FormatTime t) => t -> String

TI.formatISO8601Millis

_parseISO8601 :: (ParseTime t, Monad m, Alternative m) => String -> m t

parse an ISO8601 date-time string

Date Formatting

newtype Date

Constructors

Date

Fields

unDate :: Day

_readDate :: (ParseTime t, Monad m) => String -> m t

TI.parseTimeM True TI.defaultTimeLocale "%Y-%m-%d"

_showDate :: FormatTime t => t -> String

TI.formatTime TI.defaultTimeLocale "%Y-%m-%d"

Byte/Binary Formatting

newtype ByteArray

base64 encoded characters

Constructors

ByteArray

Fields

unByteArray :: ByteString

_readByteArray :: Monad m => Text -> m ByteArray

read base64 encoded characters

_showByteArray :: ByteArray -> Text

show base64 encoded characters

newtype Binary

any sequence of octets

Constructors

Binary

Fields

unBinary :: ByteString

_readBinaryBase64 :: Monad m => Text -> m Binary

_showBinaryBase64 :: Binary -> Text

Lens Type Aliases

type Lens_' s a = Lens_ s s a a

type Lens_ s t a b = forall (f :: * -> *). Functor f => (a -> f b) -> s -> f t